pyproject.toml
[project]
name = "parameter-golf-autoresearch"
version = "0.1.0"
description = "Autonomous ML research agent for the Parameter Golf challenge — two-tier compute with local MLX and RunPod 8xH100"
requires-python = ">=3.10"
dependencies = [
    # from openai/parameter-golf
    "numpy",
    "tqdm",
    "torch",
    "huggingface-hub",
    "kernels",
    "setuptools",
    "typing-extensions==4.15.0",
    "datasets",
    "tiktoken",
    "sentencepiece",
    # MLX (Mac only)
    "mlx>=0.22; sys_platform == 'darwin'",
    # research pipeline
    "httpx[asyncio]>=0.27",
    "feedparser>=6.0",
    "beautifulsoup4>=4.12",
    "python-dotenv>=1.0",
    "zstandard>=0.22",
    # RunPod compute layer
    "runpod>=1.6",
    "paramiko>=3.4",
    # Tavily web search + extract
    "tavily-python>=0.5",
    "apscheduler>=3.10",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
